refactor: Refactor socket handler modules
- delete the redundant `socketManager.js` module.
- move `handleConnect` functionality into `initSocket` function in `configHandlers.js`.
- change `initSocket` function name to `handleConnect` in `configHandlers.js`.
- remove `handleActivity` function in `displayHandler.js`.
- remove the direct export of `handleChatInit` function.
- remove integration of `useDispatch` and `useSelector` hooks in `chatContext.js`.
-change import names to reflect the new naming conventions in `chatContext.js`.
- remove the `handleSocketEvents` integration in `chatContext.js`.

- Note: These changes simplify socket management by consolidating functionality, removing unused code, and improving naming conventions to enhance clarity and maintainability.
