+++ frontend/src/features/socket/configHandlers.js
@@ -1,8 +1,16 @@
 import io from "socket.io-client";
+import { toast } from "react-toastify";
 
-export const handleConnect = async (url, token) => {
+// remove toast from this module
+export const initSocket = async (token, setupCallback) => {
+  if (!token) {
+    console.error("User not authorized, no token.");
+    return null;
+  }
+
   try {
-    const socket = io(url, {
+    // implement conditional API url based on process.env.mode
+    const socket = io("http://localhost:8000", {
       reconnection: false,
       withCredentials: true,
       auth: {
@@ -10,10 +18,29 @@ export const handleConnect = async (url, token) => {
       },
     });
 
+    socket.on("connect", () => {
+      setupCallback(socket);
+    });
+
+    socket.on("connect_error", (err) => {
+      console.error("Connection failed:", err.message);
+      toast.error("Unable to connect to the server. Please try again later.");
+    });
+
+    socket.on("reconnect_attempt", () => {
+      console.log("Attempting to reconnect");
+    });
+
+    socket.on("disconnect", () => {
+      console.log("Socket disconnected");
+    });
+
     return socket;
   } catch (error) {
-    console.error("Error connecting socket:", error);
-    return null;
+    console.error("Socket connection returned null or undefined.");
+    toast.error(
+      "An unexpected error occurred. Please refresh the page or try again later."
+    );
   }
 };
 
@@ -26,10 +53,3 @@ export const handleDisconnect = async (socketInstance) => {
     console.error("Error disconnecting socket:", error);
   }
 };
-
-const configAPI = {
-  handleConnect,
-  handleDisconnect,
-};
-
-export default configAPI;
